// src/app/models/jadual.model.ts
export class Jadual {
    id: string;
    anjal: string
    driverID: string;
    plateNumber: string;
    hari: string;
    tarikh: Date;
    lokasiA: string;
    lokasiB: string;
    lokasiC: string;
    timeStart: string;
    timeMiddle: string;
    timeEnd: string;
    lebihMasa: boolean;
    status: string;
}